feat(proxy-cache)_: move lua code to separate folders
