+++ nginx-proxy/lua/cache/cache.lua
