+++ nginx-proxy/lua/init_worker.lua
@@ -1,8 +1,8 @@
-local provider_loader = require("provider_loader")
+local provider_loader = require("providers.provider_loader")
 local schedule_reload_providers = provider_loader.schedule_reload_providers
 
 -- Initialize auth configuration
-local auth_config = require("auth_config")
+local auth_config = require("auth.auth_config")
 auth_config.init()
 
 -- Read URL from environment variable
