+++ nginx-proxy/nginx.conf
@@ -42,7 +42,7 @@ http {
         # Auth endpoints - no authentication required, handled by Go service
         location /auth/ {
             set_by_lua_block $go_auth_service_url {
-                local auth_config = require("auth_config")
+                local auth_config = require("auth.auth_config")
                 return auth_config.get_go_auth_service_url()
             }
             proxy_pass $go_auth_service_url;
@@ -55,14 +55,14 @@ http {
         # Internal auth endpoint for JWT token validation (hybrid Lua + Go)
         location = /_auth_token {
             internal;
-            content_by_lua_file lua/auth_token_validator.lua;
+            content_by_lua_file lua/auth/auth_token_validator.lua;
         }
                 
         # Internal endpoint for Go service validation (called by Lua)
         location = /_auth_go_verify {
             internal;
             set_by_lua_block $go_auth_service_url {
-                local auth_config = require("auth_config")
+                local auth_config = require("auth.auth_config")
                 return auth_config.get_go_auth_service_url()
             }
             proxy_pass $go_auth_service_url/auth/verify;
